pyproject.toml
[tool.poetry]
name = "hw-30-django-jwt"
version = "0.1.0"
description = ""
authors = ["KeitoAV <keito_ekka@mail.ru>"]
readme = "README.md"
packages = [{include = "hw_30_django_jwt"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
